fix: `status` 기능 변경
status 명령이 들어오면 각 진영의 점수를 출력해준다. (게임 종료 X)
