+++ README.md
@@ -32,7 +32,7 @@
 ---
 
 * [x] 킹이 잡히면 King이 잡힌 팀이 패배한다. (게임 종료)
-* [x] `status` 명령이 들어오면 각 진영의 점수를 출력해주면서 게임을 종료한다.
+* [x] `status` 명령이 들어오면 각 진영의 점수를 출력해준다.
   * [x] queen은 9점, rook은 5점, bishop은 3점, knight는 2.5점으로 계산한다.
   * [x] pawn의 기본 점수는 1점이지만 같은 file에 같은 색의 폰이 잇는 경우 0.5점으로 계산된다.
   * [x] king은 점수 계산에 포함되지 않는다.
