+++ src/main/java/chess/domain/state/Ready.java
@@ -2,8 +2,6 @@ package chess.domain.state;
 
 import chess.domain.ChessBoard;
 import chess.domain.Command;
-import chess.domain.piece.Team;
-import java.util.Map;
 
 public final class Ready implements State {
     @Override
@@ -14,11 +12,6 @@ public final class Ready implements State {
     @Override
     public State stop() {
         return new End();
-    }
-
-    @Override
-    public Map<Team, Double> status(ChessBoard chessBoard) {
-        throw new IllegalArgumentException("지금은 점수를 계산할 수 없습니다.");
     }
 
     @Override
