+++ src/main/java/chess/domain/state/White.java
@@ -21,11 +21,6 @@ public class White implements State {
     }
 
     @Override
-    public Map<Team, Double> status(ChessBoard chessBoard) {
-        throw new IllegalArgumentException("지금은 점수를 계산할 수 없습니다.");
-    }
-
-    @Override
     public State changeTurn(Command command, ChessBoard chessBoard) {
         checkTeam(command, chessBoard);
